Refactor (SPTabBar): Move new extension to correct files
