+++ solutions/devsprint-bruno-faganello-2/SpaceApp/Components/SPTabBar.swift
@@ -44,51 +44,3 @@ struct SPTabBar_Previews: PreviewProvider {
         SPTabBar()
     }
 }
-
-extension UIColor {
-    static var SP: SPUIColorExtension { .init() }
-}
-
-struct SPUIColorExtension {
-    
-    var blueSelected: UIColor { makeColor(red: 52, green: 120, blue: 246) }
-    var grayUnselected: UIColor { makeColor(red: 113, green: 113, blue: 113) }
-    var darkGray: UIColor { makeColor(red: 29, green: 29, blue: 29) }
-    
-    let contentBlack: UIColor = .black
-    
-    //MARK: - Helpers
-    private func makeColor(red: CGFloat, green: CGFloat, blue: CGFloat) -> UIColor {
-        .init(red: red/255, green: green/255, blue: blue/255, alpha: 1)
-    }
-}
-
-
-extension Color {
- 
-    func uiColor() -> UIColor {
-
-        if #available(iOS 14.0, *) {
-            return UIColor(self)
-        }
-
-        let components = self.components()
-        return UIColor(red: components.r, green: components.g, blue: components.b, alpha: components.a)
-    }
-
-    private func components() -> (r: CGFloat, g: CGFloat, b: CGFloat, a: CGFloat) {
-
-        let scanner = Scanner(string: self.description.trimmingCharacters(in: CharacterSet.alphanumerics.inverted))
-        var hexNumber: UInt64 = 0
-        var r: CGFloat = 0.0, g: CGFloat = 0.0, b: CGFloat = 0.0, a: CGFloat = 0.0
-
-        let result = scanner.scanHexInt64(&hexNumber)
-        if result {
-            r = CGFloat((hexNumber & 0xff000000) >> 24) / 255
-            g = CGFloat((hexNumber & 0x00ff0000) >> 16) / 255
-            b = CGFloat((hexNumber & 0x0000ff00) >> 8) / 255
-            a = CGFloat(hexNumber & 0x000000ff) / 255
-        }
-        return (r, g, b, a)
-    }
-}
