+++ solutions/devsprint-bruno-faganello-2/SpaceApp/Utils/Extensions/UIColor/UIColor+extensions.swift
@@ -1,0 +1,13 @@
+//
+//  UIColor+extensions.swift
+//  SpaceApp
+//
+//  Created by Paolo Prodossimo Lopes on 26/04/22.
+//
+
+import UIKit
+
+extension UIColor {
+    static var SP: SPUIColorExtension { .init() }
+}
+
